\name{armademo-package}
\alias{armademo-package}
\alias{armademo}
\docType{package}
\title{\packageTitle{armademo}}
\description{\packageDescription{armademo}}
\section{Package Content}{\packageIndices{armademo}}
\author{\packageAuthor{armademo}}
\section{Maintainer}{\packageMaintainer{armademo}}
\keyword{package}
